Improve typings further
